Plugins: split manager.go into areas of responsibility
