+++ pkg/plugins/manager/store.go
@@ -38,6 +38,50 @@ func (m *PluginManager) Plugins(_ context.Context, pluginTypes ...plugins.Type) 
 	return pluginsList
 }
 
+func (m *PluginManager) plugin(pluginID string) (*plugins.Plugin, bool) {
+	m.pluginsMu.RLock()
+	defer m.pluginsMu.RUnlock()
+	p, exists := m.store[pluginID]
+
+	if !exists || (p.IsDecommissioned()) {
+		return nil, false
+	}
+
+	return p, true
+}
+
+func (m *PluginManager) plugins() []*plugins.Plugin {
+	m.pluginsMu.RLock()
+	defer m.pluginsMu.RUnlock()
+
+	res := make([]*plugins.Plugin, 0)
+	for _, p := range m.store {
+		if !p.IsDecommissioned() {
+			res = append(res, p)
+		}
+	}
+
+	return res
+}
+
+func (m *PluginManager) isRegistered(pluginID string) bool {
+	p, exists := m.plugin(pluginID)
+	if !exists {
+		return false
+	}
+
+	return !p.IsDecommissioned()
+}
+
+func (m *PluginManager) registeredPlugins() map[string]struct{} {
+	pluginsByID := make(map[string]struct{})
+	for _, p := range m.plugins() {
+		pluginsByID[p.ID] = struct{}{}
+	}
+
+	return pluginsByID
+}
+
 func (m *PluginManager) Add(ctx context.Context, pluginID, version string) error {
 	var pluginZipURL string
